Tidied up the Animation folder. Animation Manager now in its own folder, with all methods split out to their own files. Animation Components moved to components folder. AnimationFrame renamed to Frame and moved inside the frame folder, as it's frame specific.
