+++ v3/src/animation/frame/components/Load.js
