+++ v3/src/animation/frame/components/Pause.js
