+++ v3/src/animation/frame/components/Paused.js
