+++ v3/src/animation/frame/components/Progress.js
