+++ v3/src/animation/frame/components/Stop.js
