+++ v3/src/animation/frame/components/TimeScale.js
