+++ v3/src/animation/frame/components/TotalProgress.js
