+++ v3/src/animation/frame/components/Yoyo.js
