+++ v3/src/animation/manager/AddAnimation.js
@@ -1,0 +1,16 @@
+var AddAnimation = function (key, animation)
+{
+    if (this.anims.has(key))
+    {
+        console.error('Animation with key', key, 'already exists');
+        return;
+    }
+
+    animation.key = key;
+
+    this.anims.set(key, animation);
+
+    return this;
+};
+
+module.exports = AddAnimation;
