+++ v3/src/animation/manager/GetAnimation.js
@@ -1,0 +1,6 @@
+var GetAnimation = function (key)
+{
+    return this.anims.get(key);
+};
+
+module.exports = GetAnimation;
