+++ v3/src/animation/manager/RemoveAnimation.js
@@ -1,0 +1,13 @@
+var RemoveAnimation = function (key)
+{
+    var anim = this.get(key);
+
+    if (anim)
+    {
+        this.anims.delete(key);
+    }
+
+    return anim;
+};
+
+module.exports = RemoveAnimation;
