+++ v3/src/boot/Game.js
@@ -15,7 +15,7 @@ var MainLoop = require('./MainLoop');
 var CreateRenderer = require('./CreateRenderer');
 var GlobalInputManager = require('../input/GlobalInputManager');
 var GlobalStateManager = require('../state/GlobalStateManager');
-var AnimationManager = require('../animation/AnimationManager');
+var AnimationManager = require('../animation/manager/AnimationManager');
 var TextureManager = require('../textures/TextureManager');
 var Data = require('../components/Data');
 var Cache = require('../cache/Cache');
