+++ v3/src/checksum.js
@@ -1,4 +1,4 @@
 var CHECKSUM = {
-build: '314dd020-1f74-11e7-b5ea-2d9635d37e06'
+build: 'e7a0e0b0-1f7e-11e7-9f4f-1ff609af3206'
 };
 module.exports = CHECKSUM;
